+++ src/lib.rs
@@ -152,6 +152,7 @@ fn set_x52pro_inputs_from_ship_statues(
             Attribute::HeatSink => vec![Control::HeatSink],
             Attribute::LandingGear => vec![Control::LandingGear],
             Attribute::SilentRunning => vec![Control::SilentRunning],
+            Attribute::Throttle => vec![Control::Throttle],
         }
     }